Move Jetpack connection from benefits to profiler completion
